[#S11P21S004-490] [FE] fix: pace 무한대 버그 수정

See merge request s11-s-project/S11P21S004!158

+++ android/presentation/src/main/java/com/ssafy/presentation/core/exercise/manager/ReportsManager.kt
@@ -53,7 +53,7 @@ class ReportsManager @Inject constructor(
             this.count { it.beatsPerMinute in 138..156 },
             this.count { it.beatsPerMinute in 157..176 },
             this.count { it.beatsPerMinute in 177..196 },
-        ).map { it / size * 100 }
+        ).map { (it.toDouble() / size * 100).toInt() }
 
     private val List<ExerciseSessionData>.splitData: SplitData
         get() = SplitData(
